Separates the GeneratedAttribute class out into a separate file. This will be shared across generators.
